move rpc up
