+++ cmd/cxod/cxod.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/skycoin/cxo/data"
 	"github.com/skycoin/cxo/node"
-	"github.com/skycoin/cxo/node/rpc/server"
+	"github.com/skycoin/cxo/rpc/server"
 	"github.com/skycoin/cxo/skyobject"
 )
