+++ rpc/client/example_test.go
